Mouse-Touch_Play, Playful Blob
1. I set up mouse input so that I could actually start playing with the blob. I realised that I would rather build interaction off playful mechanical explorations, as much as I enjoy aesthetics, I don't think aesthetics alone makes the kind of experience I imagine emerging.
I spent aaages trying to figure out how to deform the raymarched sphere in the hlsl code. It is kindof above me at this stage.
I bumped against limits of my understanding, but did encounter things I would like to explore further, like how to use the camera's depth and stuff like that, understanding how world space and screen space actually gets mapped out in the shader.
It's breaking my brain a bit to translate the concepts I have in my mind into the clunky lock-step sequence of the shader's logic.
I'm just trying to avoid slipping into a masochistic struggle of frustratedly working through internet posts and youtube tutorials and not really creating anything.
I set up a thing I AM familiar with, which is joining the spheres with a physics joint and now they get to interact physicslish. A quick (compromise?) to enable the playful interaction to advance.
I guess the question is, what is the value of struggling through coding struggles and learning, and what exactly does the nerdy grind produce?