A new trail spawner
Tried another approach to spawning spheres in a tail.
The calculations are neater, but still not 100%.
There is a slight jitter when the tail disappears.
It has something to do with a condition that will either spawn or delete spheres when the "tail" object passes through a treshold. I think there is some weird floating point things happening h=for a few frames in that threshold, so it might be deleting/spawning rapidly creating a little flicker.
something about that flicker is interesting, like it's the flicker that breaks the illusion and draws attention to digital processes.