New Shader
I am building a new shader from scratch, with the hope that i can build the sphere tracer a bit more "bare bones" than the one I copied-pastes from Sebastian Lague.
I am still copying a lot of code from a tutorial, and a lot of it I dont understand, not only the syntax or whatever. I concede that actually learning shaders will take a lot of time and technical understanding of rendering pipelines.
So in leu of that, is there something about the process of copying instructinos from tutorials of pasting together a frankenstein code monster to just achieve a purpose that I am happy to embrace? I'm trying to suss out how I feel about this process, like, how am I expressing agency or negotiating agency in this human-software interaction.
Situating it means I should probably also consider how i relate to the "community" or individuals that present tutorials or provide code examples. I could ask what kinds of values or purposes are included in the full assemblage of interactions, and what happens.. what can I negoiate, how does my agency get shaped in this specific situation.
I also value embodiment, and thinking about cyborg feminism (which I hold in my mind) - moving from the disembodied, the vision-oriented, towards including/centering the body(however that matters) and valuing touch as much as or more than vision...
With all of this in mind, I am struggling to see much opportunity in this activity for exploring the above? I would like to express something that matters, to be able to re-define this activity as a negotiation of agency that matters... is it possible?