tail scale falloff
staring to work on scaling the spheres according to the stretch of the chain
i still have a "flickering" when the chain moves too fast, because the positions of the spheres gets
updated 1. according to the velocity, but then 2. snapped to a (max distance) if it's too far. and each position updates independent of the other, so not necessarily in the order of the chain. so it's basically jumping between two positions from frame to frame.
i think i can fix that by smoothing the control point? and the camera?