tendril
i wanted to use the fabrik logic along with my elastic springs.
this means it solves the particle spring connections in BOTH directions,
first towards the one ahead of it, then towards the one behind it.
this means i can set the particle[0] as the anchor point.
So I wanted to make it like a tendril that extends from a first person view, kind of like an alien limb.
It wriggles about in that elastic kind of way.
Some settings can be tweaked, like the constant for the spring force, and the amount the force decays over time (its very minimal right now, so it keeps wriggling quite a bit)